+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ---
+v1.2.2
+features:
+- when an ammo type runs out, now it is possible to automatically switch to the next/first available ammo type (can be set in the MCM)
+- autolower now can be set to be ignored if the pipboy light is on
+fixes:
+- fixed potential memory leak with SwitchData
+- fixed potential CTD with the emptyMag animation event
+- fixed erroneous BCR/TR logic when switching ammo from UI menus
+---
 v1.2.1
 features:
 - support for the Anniversary Update